+++ render/d3d12/d3d12renderer.h
@@ -51,7 +51,7 @@ namespace Lightning
 			IRenderFence* CreateRenderFence()override;
 			IDevice* CreateDevice()override;
 			ISwapChain* CreateSwapChain()override;
-			IDepthStencilBuffer* CreateDepthStencilBuffer(std::size_t width, std::size_t height)override;
+			IDepthStencilBuffer* CreateDepthStencilBuffer(std::uint32_t width, std::uint32_t height)override;
 			RenderPass* CreateRenderPass(RenderPassType type)override;
 		private:
 			struct PipelineStateRootSignature
